can收发配置修改

+++ Main-Ctrl/H7_CTRL-v5/Bsp/inc/can_bsp.h
@@ -7,8 +7,8 @@
 
 void bsp_can_init(void);
 void can_filter_init(void);
-uint8_t fdcanx_send_data(hcan_t *hfdcan, uint16_t id, uint8_t *data, uint32_t len);
-uint8_t fdcanx_receive(hcan_t *hfdcan, uint16_t *rec_id, uint8_t *buf);
+uint8_t fdcanx_send_data(hcan_t *hfdcan, uint32_t id, uint8_t *data, uint32_t len);
+uint8_t fdcanx_receive(hcan_t *hfdcan, uint32_t *rec_id, uint8_t *buf);
 void fdcan1_rx_callback(void);
 void fdcan2_rx_callback(void);
 void fdcan3_rx_callback(void);
